Trace when received index response (only those with OK status)
Convert raw cycles from packet index to normal timestamps to be logged
in tracepoint.

Use stream id from request, as reply message appears to have that always
set to zero.

As a quick hack to avoid messing with autotools, check-in
lttng-live-trace.{h,c} generated files and explicitly and
unconditionally link to liblttng-ust.