Live: comm: bundle TCP socket write write sequence in a single write
Delayed ack induced delay were observed [1] on the second write of the
"write cmd, write request" sequence and resulted in non-negligible
latency of lttng-live. This is due to the combination of the
Nagle's algorithm and the delayed ack mechanism. The write write read
sequence is a worst-case when both are activated. The second write is
held up until the first write is acknowledged (via ack). The solution to
this is to bundle the write action into a single one [2].

[1] https://github.com/tbricks/wireshark-lttng-plugin
    Basic Wireshark dissector for lttng-live by Anto Smyk from Itiviti
[2] https://lists.freebsd.org/pipermail/freebsd-net/2006-January/009527.html

Reported-and-tested-by: Anton Smyk <anton.smyk@itiviti.com>
Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>