Add basic, LTTng-independent "list sessions" `src.ctf.lttng-live` test
This patch adds a very basic test for the "list sessions" feature of
`src.ctf.lttng-live`. The goal of this test is to be a starting point
for the development of future, more specific `src.ctf.lttng-live` tests.

Because `src.ctf.lttng-live` needs a running LTTng relay daemon to
connect to and because we don't want the Babeltrace tests to depend on
LTTng, the new `lttng_live_list_sessions_faux_server.py` script mimics
an LTTng relay daemon for the "list sessions" use case. The script
expects exact commands from an LTTng live viewer and replies with exact
responses. The actual expected and sent bytes were directly copied from
a TCP conversation captured with Wireshark.

The faux server script accepts two arguments: the first is the name of a
temporary file where to write the server's TCP port number; the second
is the name of another temporary file to which to rename the first one
once the port number is written (synchronization point for the shell
test script). Because we want multiple instances of this test to be able
to execute at the same time, the faux server does not use a fixed port
number: it binds to port 0 so that the OS decides which free TCP port to
use.

The test script, `tests/plugins/src.ctf.lttng-live/test_succeed`, does
this:

1. It runs the faux server, passing two temporary file names, and making
   it a background job.

2. It waits for the second temporary file to contain anything.

3. It reads the port number from the second temporary file.

4. It runs:

       $ babeltrace2 -i lttng-live net://localhost:PORT

   where `PORT` is the faux server's port number. The CLI should list
   the available sessions with this command line format.

5. It compares the output of 4. to an expected file generated from a
   template where the port number is variable (as the CLI's output
   contains that port number).

6. It removes the temporary and expectation files.

The faux server exits itself.

In the test script, the test_list_sessions() function is very specific
to this use case. Eventually, the goal is to make it more generic,
accepting a faux server script path and an expectation file path, for
example.

This test only runs if you have a Python interpreter (`HAVE_PYTHON`
Automake definition).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Iad1a46995eacdc83c58c45746cde3f9980b10a02